requirements.txt
tensorflow
pillow
scipy
matplotlib